Rename AIBackend models and fix namespacing
