+++ app/services/ai_backend.rb
@@ -1,0 +1,4 @@
+class AIBackend
+  attr :client
+
+end
