+++ app/services/ai_backend/open_ai.rb
@@ -1,25 +1,23 @@
-class AIBackends::Anthropic
-  attr :client
-
+class AIBackend::OpenAI < AIBackend
   # Rails system tests don't seem to allow mocking because the server and the
   # test are in separate processes.
   #
-  # In regular tests, mock this method or the TestClients::Anthropic class to do
+  # In regular tests, mock this method or the TestClient::OpenAI class to do
   # what you want instead.
   def self.client
     if Rails.env.test?
-      TestClients::Anthropic
+      ::TestClient::OpenAI
     else
-      Anthropic::Client
+      ::OpenAI::Client
     end
   end
 
   def initialize(user, assistant, conversation, message)
-    raise Anthropic::ConfigurationError if user.anthropic_key.blank?
+    raise ::OpenAI::ConfigurationError if user.openai_key.blank?
     begin
-      @client = self.class.client.new(access_token: user.anthropic_key)
+      @client = self.class.client.new(access_token: user.openai_key)
     rescue ::Faraday::UnauthorizedError => e
-      raise Anthropic::ConfigurationError
+      raise ::OpenAI::ConfigurationError
     end
     @assistant = assistant
     @conversation = conversation
@@ -30,7 +28,7 @@ class AIBackends::Anthropic
     stream_response_text = ""
 
     response_handler = proc do |intermediate_response, bytesize|
-      chunk = intermediate_response.dig("delta", "text")
+      chunk = intermediate_response.dig("choices", 0, "delta", "content")
       print chunk if Rails.env.development?
       if chunk
         stream_response_text += chunk
@@ -39,30 +37,27 @@ class AIBackends::Anthropic
     rescue ::GetNextAIMessageJob::ResponseCancelled => e
       raise e
     rescue ::Faraday::UnauthorizedError => e
-      raise Anthropic::ConfigurationError
+      raise ::OpenAI::ConfigurationError
     rescue => e
-      puts "\nUnhandled error in AIBackends::Anthropic response handler: #{e.message}"
+      puts "\nUnhandled error in AIBackend::OpenAI response handler: #{e.message}"
       puts e.backtrace
     end
 
     response_handler = nil unless block_given?
 
     begin
-      response = @client.messages(
+      response = @client.chat(parameters: {
         model: @assistant.model,
-        system: @assistant.instructions,
-        messages: preceding_messages,
-        parameters: {
-          max_tokens: 2000, # we should really set this dynamically, based on the model, to the max
-          stream: response_handler,
-        }
-      )
+        messages: system_message + preceding_messages,
+        stream: response_handler,
+        max_tokens: 2000, # we should really set this dynamically, based on the model, to the max
+      })
     rescue ::Faraday::UnauthorizedError => e
-      raise Anthropic::ConfigurationError
+      raise ::OpenAI::ConfigurationError
     end
 
-    response_text = if response.is_a?(Hash) && response.dig("content")
-      response.dig("content", 0, "text")
+    response_text = if response.is_a?(Hash) && response.dig("choices")
+      response.dig("choices", 0, "message", "content")
     else
       response
     end
@@ -76,19 +71,22 @@ class AIBackends::Anthropic
 
   private
 
+  def system_message
+    return [] if @assistant.instructions.blank?
+
+    [{
+      role: 'system',
+      content: @assistant.instructions
+    }]
+  end
+
   def preceding_messages
     @conversation.messages.for_conversation_version(@message.version).where("messages.index < ?", @message.index).collect do |message|
       if @assistant.images && message.documents.present?
 
         content = [{ type: "text", text: message.content_text }]
         content += message.documents.collect do |document|
-          { type: "image",
-            source: {
-              type: "base64",
-              media_type: document.file.blob.content_type,
-              data: document.file_base64(:large),
-            }
-          }
+          { type: "image_url", image_url: { url: document.file_data_url(:large) }}
         end
 
         {
