+++ test/jobs/get_next_ai_message_job_anthropic_test.rb
@@ -6,7 +6,7 @@ class GetNextAIMessageJobAnthropicTest < ActiveJob::TestCase
     @user = @conversation.user
     @conversation.messages.create! role: :user, content_text: "Still there?", assistant: @conversation.assistant
     @message = @conversation.latest_message_for_version(:latest)
-    @test_client = TestClients::Anthropic.new(access_token: 'abc')
+    @test_client = TestClient::Anthropic.new(access_token: 'abc')
   end
 
   test "populates the latest message from the assistant" do
@@ -45,7 +45,7 @@ class GetNextAIMessageJobAnthropicTest < ActiveJob::TestCase
   end
 
   test "when API response key is, a nice error message is displayed" do
-    TestClients::Anthropic.stub :text, "" do
+    TestClient::Anthropic.stub :text, "" do
       assert GetNextAIMessageJob.perform_now(@user.id, @message.id, @conversation.assistant.id)
       assert_includes @conversation.latest_message_for_version(:latest).content_text, "a blank response"
     end
